requirements.txt
flask
couchbase
python-dotenv